Restore green main: rustfmt and source-file-length policy (#545)
Two portable gates were failing on main after the #535 merge, and because
the windows_native job runs those same checks before any native step,
they were suppressing all Windows validation on the default branch.

Apply rustfmt to src/runtime/agent_probe.rs, whose retry-loop Err arm
reflowed after the loader-transient change.

Split tests/issue382_behavior.rs, which reached 1028 lines against the
1000-line hard limit. The CW02-07 remote-plan golden helpers move to
tests/issue382/remote_plan_golden.rs, matching the existing per-concern
helper layout under tests/issue382/. Only helpers move; every #[test]
entry point stays put and all 28 tests still pass, so acceptance
semantics are unchanged.

This is issue #545 deliverable 5 and a precondition for V7.
